Move subcomponent to separate file
